fix(ci): silence clang-tidy sign-conversion warning in counter draw
clang-tidy 22 (clang-pr CI lane) reported:

  chronoid/uuidv7/uuidv7_sequence.c:54:48: warning: implicit conversion
  changes signedness: 'int' to 'unsigned int' [clang-diagnostic-sign-conversion]

The expression was:

  *out = (uint16_t) ((((uint16_t) two[0] << 8) | two[1]) & UUIDV7_COUNTER_MASK);

C integer promotion takes the (uint16_t) two[0] value to int before the
shift; the |two[1] keeps it int; then & UUIDV7_COUNTER_MASK (which is
0x0FFFu — unsigned int) drove an int -> unsigned int implicit
conversion that clang-tidy's sign-conversion check rejects.

Fix: cast the high byte to unsigned int (not uint16_t) before the
shift, so the entire arithmetic stays in unsigned-int land. Behavior
is unchanged — UUIDV7_COUNTER_MASK = 0x0FFF clamps the result to 12
bits regardless of the wider intermediate type, and the final
(uint16_t) cast on the assignment side has the same observable
result. Pure compile-time-warning fix; no runtime change.

Refs: PR #7 (round 4)

+++ chronoid/uuidv7/uuidv7_sequence.c
@@ -51,7 +51,16 @@ chronoid_uuidv7_sequence_draw_counter (uint16_t *out)
       chronoid_uuidv7_sequence_draw_random (two, sizeof two);
   if (e != CHRONOID_UUIDV7_OK)
     return e;
-  *out = (uint16_t) ((((uint16_t) two[0] << 8) | two[1]) & UUIDV7_COUNTER_MASK);
+  /* Cast the high byte to unsigned int (not uint16_t) before the
+   * shift so the entire AND with UUIDV7_COUNTER_MASK (which is
+   * `unsigned int` per the 'u' suffix on 0x0FFFu) stays in
+   * unsigned-int land throughout. Without this cast the C integer
+   * promotion rule turns the shifted value into `int`, and the
+   * subsequent AND triggers an implicit int -> unsigned int
+   * conversion that clang-tidy's sign-conversion check rejects. */
+  *out =
+      (uint16_t) ((((unsigned int) two[0] << 8) | two[1]) &
+      UUIDV7_COUNTER_MASK);
   return CHRONOID_UUIDV7_OK;
 }
